requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
scrapy==2.11.0
aiohttp==3.9.1
python-dateutil==2.8.2
feedparser==6.0.10
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
psycopg2-binary==2.9.9
